reorganize files and setup to use context
