--- packages/app-shell/src/PlanSelector/components/PlanSelectorParent.stories.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import { PlanSelectorParent } from './PlanSelectorParent';
-import response from '../mock';
-
-export default {
-  title: 'Plan Selector Parent',
-  component: PlanSelectorParent,
-};
-
-const Template = (args) => <PlanSelectorParent {...args} />;
-
-export const AnotherExample = Template.bind({});
-AnotherExample.args = {
-  planOptions:
-    response.data.account.currentOrganization.billing.changePlanOptions,
-};
